Fix compilation on Windows
