+++ comp/core/workloadmeta/collectors/internal/remote/processcollector/process_collector.go
@@ -12,6 +12,7 @@ package processcollector
 import (
 	"context"
 	"errors"
+	"fmt"
 	"strconv"
 	"time"
